Service_FastAPI/requirements.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
fastapi==0.79.0
h11==0.13.0
idna==3.3
pip==1.9.1
requests==2.28.1
setuptools==60.2.0
urllib3==1.26.11
wheel==0.37.1
pydantic==1.9.1
python-dotenv==0.20.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
python-dotenv==0.20.0
